Fixed materials amount.

+++ JOMonitoringApp/Views/Reports/JobOrderEstimateOfMaterials/frmEstimateOfMaterials.cs
@@ -95,7 +95,7 @@ namespace JOMonitoringApp.Views.Reports
                 newRow["item"] = dataRow["item_name"];
                 newRow["qty"] = dataRow["item_quantity"];
                 newRow["unit_cost"] = dataRow["item_quantity"] != DBNull.Value && Convert.ToDecimal(dataRow["item_quantity"]) != 0
-                    ? (Convert.ToDecimal(dataRow["item_cost"]) / Convert.ToDecimal(dataRow["item_quantity"])).ToString()
+                    ? (Convert.ToDouble(dataRow["item_cost"]) / Convert.ToDouble(dataRow["item_quantity"])).ToString()
                     : "0";
                 newRow["amount"] = Convert.ToDecimal(dataRow["item_cost"]);
